fix linter

+++ pumpkin/src/net/packet/play.rs
@@ -434,7 +434,7 @@ impl Player {
         let mut inventory = self.inventory().lock().await;
 
         let source_slot = inventory.get_slot_with_item(block.item_id);
-        let mut dest_slot = inventory.get_empty_hotbar_slot() as usize;
+        let mut dest_slot = inventory.get_empty_hotbar_slot();
 
         let dest_slot_data = match inventory.get_slot(dest_slot + SLOT_HOTBAR_START) {
             Ok(Some(stack)) => stack.clone(),
